refactor: split god modules into focused sibling modules
Verbatim extractions, no behavior change:
- hooks/useAgentSession (3442->2998) + stream/queue/notices helpers
- SessionSidebar (3112->1599) + helpers/chrome/rows
- ChatInput (2814->2560) + model-options/slash-commands/draft-attachments/banners
- AppShell (2436->2283) + app-update/provider-usage/layout
- ModelsConfig (2385->1736) + types/panels
- MessageView (1827->1409) + tool-format/task-panel/diff-view
Contract test points the archive flow at SessionSidebar-rows.
