+++ components/MessageView.tsx
@@ -9,10 +9,11 @@ import { parseCompactionSummary } from "@/lib/compaction-summary";
 import { isEmptyThinkingBlock } from "@/lib/message-display";
 import { Tooltip, Collapsible, CollapsibleTrigger } from "./ui/primitives";
 import { useCopyFeedback } from "@/hooks/useCopyFeedback";
-import { SubagentStatusIcon } from "./SubagentStatusIcon";
-import { formatCost, formatDuration, formatTokens, shortModel } from "@/lib/subagent-format";
-import { parseUnifiedPatch, type SplitDiffCell } from "@/lib/patch";
 import { formatCompactNumber } from "@/lib/format";
+import { TaskResultPanel } from "./MessageView-task-panel";
+import { getResultDiff, PairedDiffResult, PairedResult } from "./MessageView-diff-view";
+import { getToolPreview, formatToolCommand, formatToolOutput, getToolResultMeta } from "./MessageView-tool-format";
+export { TaskResultPanel } from "./MessageView-task-panel";
 import type {
   AgentMessage,
   UserMessage,
@@ -801,358 +802,6 @@ const ToolCallBlock = memo(function ToolCallBlock({ block, result, duration, isS
 ));
 
 
-type TaskResultRowLike = Record<string, unknown>;
-
-function taskRowStatus(row: TaskResultRowLike): "started" | "completed" | "failed" | "aborted" {
-  if (row.aborted === true) return "aborted";
-  if (typeof row.error === "string" && row.error) return "failed";
-  if (typeof row.exitCode === "number") return row.exitCode === 0 ? "completed" : "failed";
-  const status = row.status;
-  if (status === "completed") return "completed";
-  if (status === "failed") return "failed";
-  if (status === "aborted") return "aborted";
-  return "started";
-}
-
-function TaskResultStatusIcon({ status }: { status: "started" | "completed" | "failed" | "aborted" }) {
-  return <SubagentStatusIcon status={status} />;
-}
-
-/**
- * Compact per-subagent summary rendered inside an expanded `task` tool call.
- * Feeds off the size-bounded task details allowlisted by the session reader
- * (lib/session-reader.ts stripToolResultDetails): settled results when
- * present, otherwise the mid-run progress snapshot.
- */
-function isRecord(value: unknown): value is Record<string, unknown> {
-  return typeof value === "object" && value !== null && !Array.isArray(value);
-}
-export function TaskResultPanel({ details }: { details: unknown }) {
-  const { t, tn } = useI18n();
-  if (!isRecord(details)) return null;
-  const results = (Array.isArray(details.results) ? details.results : []).filter(isRecord);
-  const progress = (Array.isArray(details.progress) ? details.progress : []).filter(isRecord);
-  const asyncInfo = isRecord(details.async) ? details.async : null;
-  if (results.length === 0 && progress.length === 0 && !asyncInfo) return null;
-
-  // Settled results win; otherwise the mid-run progress snapshot; a bare
-  // async marker (spawn recorded, no rows yet) still names the job.
-  const rows = results.length > 0
-    ? results
-    : progress.length > 0
-      ? progress
-      : asyncInfo && typeof asyncInfo.jobId === "string"
-        ? [{ id: asyncInfo.jobId, agent: "task", status: "started", task: asyncInfo.jobId } as TaskResultRowLike]
-        : [];
-  const totalTokens = rows.reduce((sum, row) => sum + (typeof row.tokens === "number" ? row.tokens : 0), 0);
-  const totalCost = rows.reduce((sum, row) => sum + (typeof row.cost === "number" ? row.cost : 0), 0);
-  const totalDurationMs = typeof details.totalDurationMs === "number" ? details.totalDurationMs : undefined;
-  const totalTokensLabel = formatTokens(totalTokens);
-  const totalParts = [
-    tn("chatWindow.subagentCount", rows.length),
-    totalTokensLabel ? t("chatWindow.tokensUnit", { count: totalTokensLabel }) : null,
-    formatCost(totalCost),
-    formatDuration(totalDurationMs),
-  ].filter(Boolean);
-
-  return (
-    <div
-      style={{
-        borderTop: "1px solid var(--border)",
-        background: "var(--bg-subtle)",
-        padding: "8px 10px",
-        display: "grid",
-        gap: 4,
-      }}
-    >
-      <div style={{ display: "flex", alignItems: "center", gap: 8, fontSize: 11, color: "var(--text-muted)" }}>
-        <span style={{ fontWeight: 600, color: "var(--text)" }}>{t("messageView.taskSubagents")}</span>
-        <span style={{ marginLeft: "auto", fontFamily: "var(--font-mono)", color: "var(--text-dim)", fontSize: 10.5 }}>
-          {totalParts.join(" · ")}
-        </span>
-        {asyncInfo && (
-          <span style={{ fontFamily: "var(--font-mono)", fontSize: 10, color: "var(--text-dim)" }}>⤴</span>
-        )}
-      </div>
-      {rows.map((row, index) => {
-        const id = typeof row.id === "string" ? row.id : `row-${index}`;
-        const status = taskRowStatus(row);
-        const task = typeof row.task === "string" && row.task ? row.task : (typeof row.assignment === "string" ? row.assignment : null);
-        const rowTokens = formatTokens(typeof row.tokens === "number" ? row.tokens : undefined);
-        const rowParts = [
-          rowTokens ? t("chatWindow.tokensUnit", { count: rowTokens }) : null,
-          formatCost(typeof row.cost === "number" ? row.cost : undefined),
-          status !== "started" ? formatDuration(typeof row.durationMs === "number" ? row.durationMs : undefined) : null,
-          shortModel(typeof row.resolvedModel === "string" ? row.resolvedModel : undefined),
-        ].filter(Boolean);
-        return (
-          <div
-            key={id}
-            aria-label={`${typeof row.agent === "string" ? row.agent : "subagent"}: ${t(`chatWindow.subagentState.${status}`)}${task ? ` — ${task}` : ""}`}
-            style={{ display: "flex", alignItems: "center", gap: 6, minWidth: 0, fontSize: 11.5 }}
-          >
-            <TaskResultStatusIcon status={status} />
-            <span style={{ fontFamily: "var(--font-mono)", fontWeight: 600, fontSize: 10.5, color: "var(--accent)", flexShrink: 0 }}>
-              {typeof row.agent === "string" ? row.agent : "subagent"}
-            </span>
-            <span style={{ minWidth: 0, overflow: "hidden", textOverflow: "ellipsis", whiteSpace: "nowrap", flex: 1, color: "var(--text)" }}>
-              {task ?? ""}
-            </span>
-            {rowParts.length > 0 && (
-              <span style={{ flexShrink: 0, fontFamily: "var(--font-mono)", fontSize: 10, color: "var(--text-dim)" }}>
-                {rowParts.join(" · ")}
-              </span>
-            )}
-          </div>
-        );
-      })}
-    </div>
-  );
-}
-
-interface ResultDiff {
-  text: string;
-}
-
-function getResultDiff(result: ToolResultMessage): ResultDiff | null {
-  const details = (result as ToolResultMessage & { details?: unknown }).details;
-  if (typeof details !== "object" || details === null || Array.isArray(details)) return null;
-  const record = details as Record<string, unknown>;
-  const patch = typeof record.patch === "string" ? record.patch : null;
-  if (patch) return { text: patch };
-  const diff = typeof record.diff === "string" ? record.diff : null;
-  if (diff) return { text: diff };
-  return null;
-}
-
-function PairedDiffResult({ diff }: { diff: ResultDiff }) {
-  return (
-    <div
-      style={{
-        borderTop: "1px solid color-mix(in srgb, var(--status-success) 15%, transparent)",
-        background: "var(--bg)",
-      }}
-    >
-      <SplitPatchView text={diff.text} />
-    </div>
-  );
-}
-
-function SplitPatchView({ text }: { text: string }) {
-  const { t } = useI18n();
-  const files = useMemo(() => parseUnifiedPatch(text), [text]);
-  if (!files) return <PatchTextView text={text} />;
-  const showFileHeaders = files.length > 1;
-
-  return (
-    <div style={{ maxHeight: 560, overflowY: "auto", overflowX: "hidden", background: "var(--bg)" }}>
-      {files.map((file, fileIndex) => (
-        <div
-          key={fileIndex}
-          style={{
-            minWidth: 0,
-            borderTop: fileIndex === 0 ? "none" : "1px solid var(--border)",
-            fontFamily: "var(--font-mono)",
-            fontSize: 12,
-            lineHeight: 1.55,
-          }}
-        >
-          {showFileHeaders && (
-            <div
-              style={{
-                display: "grid",
-                gridTemplateColumns: "minmax(0, 1fr) minmax(0, 1fr)",
-                position: "sticky",
-                top: 0,
-                zIndex: 1,
-                background: "var(--bg-panel)",
-                borderBottom: "1px solid var(--border)",
-              }}
-            >
-              <SplitDiffHeader title={file.oldPath || t("messageView.diffBefore")} side="left" />
-              <SplitDiffHeader title={file.newPath || t("messageView.diffAfter")} side="right" />
-            </div>
-          )}
-
-          <div style={{ display: "grid", gridTemplateColumns: "minmax(0, 1fr) minmax(0, 1fr)" }}>
-            {file.rows.map((row, rowIndex) => {
-              if (row.type === "hunk") {
-                return null;
-              }
-
-              return (
-                <div key={rowIndex} style={{ display: "contents" }}>
-                  <SplitDiffCellView cell={row.left} side="left" />
-                  <SplitDiffCellView cell={row.right} side="right" />
-                </div>
-              );
-            })}
-          </div>
-        </div>
-      ))}
-    </div>
-  );
-}
-
-function SplitDiffHeader({ title, side }: { title: string; side: "left" | "right" }) {
-  return (
-    <div
-      title={title}
-      style={{
-        padding: "5px 10px",
-        color: "var(--text-dim)",
-        borderRight: side === "left" ? "1px solid var(--border)" : "none",
-        overflow: "hidden",
-        textOverflow: "ellipsis",
-        whiteSpace: "nowrap",
-      }}
-    >
-      {title}
-    </div>
-  );
-}
-
-function SplitDiffCellView({ cell, side }: { cell: SplitDiffCell; side: "left" | "right" }) {
-  const bg =
-    cell.type === "added"
-      ? "color-mix(in srgb, var(--status-success) 12%, transparent)"
-      : cell.type === "removed"
-      ? "color-mix(in srgb, var(--status-error) 13%, transparent)"
-      : cell.type === "empty"
-      ? "var(--bg-subtle)"
-      : "transparent";
-  const marker =
-    cell.type === "added" ? "+" : cell.type === "removed" ? "-" : " ";
-  const markerColor =
-    cell.type === "added" ? "var(--status-success)" : cell.type === "removed" ? "var(--status-error)" : "var(--text-dim)";
-
-  return (
-    <div
-      style={{
-        display: "flex",
-        minWidth: 0,
-        background: bg,
-        borderRight: side === "left" ? "1px solid var(--border)" : "none",
-      }}
-    >
-      <span
-        style={{
-          width: 42,
-          padding: "0 6px",
-          textAlign: "right",
-          color: "var(--text-dim)",
-          userSelect: "none",
-          background: "var(--bg-panel)",
-          borderRight: "1px solid var(--border)",
-          flexShrink: 0,
-        }}
-      >
-        {cell.lineNo ?? ""}
-      </span>
-      <span
-        style={{
-          width: 18,
-          padding: "0 5px",
-          color: markerColor,
-          userSelect: "none",
-          fontWeight: cell.type === "context" || cell.type === "empty" ? 400 : 700,
-          flexShrink: 0,
-        }}
-      >
-        {marker}
-      </span>
-      <span
-        style={{
-          flex: 1,
-          minWidth: 0,
-          padding: "0 10px 0 0",
-          color: cell.type === "empty" ? "var(--text-dim)" : "var(--text)",
-          whiteSpace: "pre-wrap",
-          overflowWrap: "anywhere",
-        }}
-      >
-        {cell.text || "\u00a0"}
-      </span>
-    </div>
-  );
-}
-
-function PatchTextView({ text }: { text: string }) {
-  const lines = text.split(/\r?\n/);
-
-  return (
-    <div style={{ maxHeight: 520, overflowY: "auto", overflowX: "hidden", fontFamily: "var(--font-mono)", fontSize: 12, lineHeight: 1.55, minWidth: 0 }}>
-      {lines.map((line, i) => {
-        const kind =
-          line.startsWith("@@") ? "hunk" :
-          line.startsWith("+") && !line.startsWith("+++") ? "added" :
-          line.startsWith("-") && !line.startsWith("---") ? "removed" :
-          "context";
-        const bg =
-          kind === "added" ? "color-mix(in srgb, var(--status-success) 12%, transparent)" :
-          kind === "removed" ? "color-mix(in srgb, var(--status-error) 13%, transparent)" :
-          kind === "hunk" ? "color-mix(in srgb, var(--accent) 12%, transparent)" :
-          "transparent";
-        const color =
-          kind === "added" ? "var(--status-success)" :
-          kind === "removed" ? "var(--status-error)" :
-          kind === "hunk" ? "var(--accent)" :
-          "var(--text)";
-
-        return (
-          <div
-            key={i}
-            style={{
-              display: "flex",
-              background: bg,
-              borderLeft: kind === "added"
-                ? "3px solid var(--status-success)"
-                : kind === "removed"
-                ? "3px solid var(--status-error)"
-                : kind === "hunk"
-                ? "3px solid var(--accent)"
-                : "3px solid transparent",
-            }}
-          >
-            <span
-              style={{
-                width: 48,
-                padding: "0 8px",
-                color: "var(--text-dim)",
-                background: "var(--bg-panel)",
-                borderRight: "1px solid var(--border)",
-                textAlign: "right",
-                userSelect: "none",
-                flexShrink: 0,
-              }}
-            >
-              {i + 1}
-            </span>
-            <span style={{ padding: "0 10px", whiteSpace: "pre-wrap", overflowWrap: "anywhere", color }}>
-              {line || "\u00a0"}
-            </span>
-          </div>
-        );
-      })}
-    </div>
-  );
-}
-
-function PairedResult({ text, isEmpty, isError }: {
-  text: string;
-  isEmpty: boolean;
-  isError: boolean;
-}) {
-  const { t } = useI18n();
-  return (
-    <div className={`tool-call-output${isError ? " tool-call-output-error" : ""}`}>
-      <pre className="tool-call-output-text" data-tool-output="true">
-        {isEmpty ? t("messageView.noOutput") : text}
-      </pre>
-    </div>
-  );
-}
-
 function CompactionMessageView({ message }: { message: CustomMessage }) {
   const { t, locale } = useI18n();
   const summary = getMessageText(message.content);
@@ -1663,73 +1312,6 @@ function previewText(text: string): string {
   const normalized = text.replace(/\s+/g, " ").trim();
   if (!normalized) return translate("messageView.showExtensionMessage");
   return normalized.length > 140 ? `${normalized.slice(0, 140)}...` : normalized;
-}
-
-
-function getToolPreview(block: ToolCallContent): string {
-  const input = block.input;
-  if (!input || typeof input !== "object") return "";
-  const keys = Object.keys(input);
-  if (keys.length === 0) return "";
-
-  // Common tool input patterns
-  if ("command" in input) return String(input.command).slice(0, 120);
-  if ("path" in input) return String(input.path).slice(0, 120);
-  if ("file_path" in input) return String(input.file_path).slice(0, 120);
-  if ("pattern" in input) return String(input.pattern).slice(0, 120);
-  if ("query" in input) return String(input.query).slice(0, 120);
-
-  const first = input[keys[0]];
-  return String(first).slice(0, 120);
-}
-function formatToolCommand(block: ToolCallContent): string {
-  const input = block.input;
-  if (input && typeof input.command === "string") return input.command;
-  if (input && typeof input.path === "string") return `${block.toolName} ${input.path}`;
-  if (input && typeof input.file_path === "string") return `${block.toolName} ${input.file_path}`;
-  if (input && typeof input.query === "string") return `${block.toolName} ${input.query}`;
-  try {
-    return `${block.toolName} ${JSON.stringify(input)}`;
-  } catch {
-    return block.toolName;
-  }
-}
-
-function formatToolOutput(text: string, toolName: string): string {
-  if (!isReadToolName(toolName)) return text;
-  return text
-    .split("\n")
-    .map((line) => line.replace(/^\s*\d+:\s?/, ""))
-    .join("\n");
-}
-
-function isReadToolName(toolName: string): boolean {
-  const name = toolName.toLowerCase();
-  return name === "read" || name.endsWith(".read") || name.endsWith("_read");
-}
-
-function getToolResultMeta(result: ToolResultMessage | undefined): string | null {
-  if (!result || !isRecord(result.details)) return null;
-  const details = result.details;
-  const usage = isRecord(details.usage) ? details.usage : details;
-  const readNumber = (...keys: string[]): number | undefined => {
-    for (const key of keys) {
-      const value = usage[key];
-      if (typeof value === "number" && Number.isFinite(value) && value > 0) return value;
-    }
-    return undefined;
-  };
-  const input = readNumber("input", "inputTokens", "input_tokens");
-  const output = readNumber("output", "outputTokens", "output_tokens");
-  const cacheRead = readNumber("cacheRead", "cache_read", "cacheReadTokens");
-  const cacheWrite = readNumber("cacheWrite", "cache_write", "cacheWriteTokens");
-  const parts = [
-    input ? `in ${formatCompactNumber(input)}` : null,
-    output ? `out ${formatCompactNumber(output)}` : null,
-    cacheRead ? `cache R ${formatCompactNumber(cacheRead)}` : null,
-    cacheWrite ? `cache W ${formatCompactNumber(cacheWrite)}` : null,
-  ].filter((part): part is string => Boolean(part));
-  return parts.length > 0 ? parts.join(" · ") : null;
 }
 
 function BashExecutionView({ message, sessionId }: { message: BashExecutionMessage; sessionId?: string }) {
