+++ hooks/useAgentSession-notices.ts
@@ -1,0 +1,79 @@
+// Notice-queue state (types, reducer, ids) extracted from useAgentSession
+// (pure logic only — no hook state).
+
+export type NoticeType = "info" | "success" | "warning" | "error";
+
+export type NoticeItem = {
+  id: string;
+  message: string;
+  type: NoticeType;
+  exiting?: boolean;
+};
+
+export type NoticeState = {
+  visible: NoticeItem[];
+  pending: NoticeItem[];
+};
+
+export type NoticeAction =
+  | { type: "add"; notice: NoticeItem }
+  | { type: "mark_oldest_exiting" }
+  | { type: "remove"; id: string };
+
+export const MAX_NOTICES = 5;
+export const NOTICE_VISIBLE_MS = 5000;
+export const NOTICE_ERROR_VISIBLE_MS = 30000;
+export const NOTICE_EXIT_ANIMATION_MS = 180;
+
+export function createNoticeId(): string {
+  if (typeof crypto !== "undefined" && typeof crypto.randomUUID === "function") {
+    return crypto.randomUUID();
+  }
+  return `${Date.now()}-${Math.random().toString(36).slice(2)}`;
+}
+
+export function markOldestNoticeExiting(notices: NoticeItem[]): NoticeItem[] {
+  const index = notices.findIndex((notice) => !notice.exiting);
+  if (index === -1) return notices;
+  return notices.map((notice, i) => (
+    i === index ? { ...notice, exiting: true } : notice
+  ));
+}
+
+export function fillPendingNotices(visible: NoticeItem[], pending: NoticeItem[]): NoticeState {
+  let nextVisible = visible;
+  let nextPending = pending;
+  while (nextPending.length > 0 && nextVisible.length < MAX_NOTICES) {
+    const [next, ...rest] = nextPending;
+    nextVisible = [...nextVisible, next];
+    nextPending = rest;
+  }
+  if (nextPending.length > 0 && !nextVisible.some((notice) => notice.exiting)) {
+    nextVisible = markOldestNoticeExiting(nextVisible);
+  }
+  return { visible: nextVisible, pending: nextPending };
+}
+
+export function noticeReducer(state: NoticeState, action: NoticeAction): NoticeState {
+  switch (action.type) {
+    case "add": {
+      if (state.visible.some((notice) => notice.exiting) || state.visible.length >= MAX_NOTICES) {
+        return {
+          visible: state.visible.some((notice) => notice.exiting)
+            ? state.visible
+            : markOldestNoticeExiting(state.visible),
+          pending: [...state.pending, action.notice],
+        };
+      }
+      return { ...state, visible: [...state.visible, action.notice] };
+    }
+    case "mark_oldest_exiting":
+      return { ...state, visible: markOldestNoticeExiting(state.visible) };
+    case "remove": {
+      const visible = state.visible.filter((notice) => notice.id !== action.id);
+      return fillPendingNotices(visible, state.pending);
+    }
+    default:
+      return state;
+  }
+}
