+++ lib/api-contract.test.mjs
@@ -37,9 +37,11 @@ test("session archive route stops live children and maps missing sessions", asyn
 
 test("session archive remains keyboard-discoverable with an ARIA label", async () => {
   const source = await readFile(new URL("../components/SessionSidebar.tsx", import.meta.url), "utf8");
-  assert.match(source, /api\/sessions\/\$\{encodeURIComponent\(session\.id\)\}\/archive/);
-  assert.match(source, /sessionSidebar\.archiveLeafOnly/);
-  assert.match(source, /sessionSidebar\.archiveConfirm/);
+  const rows = await readFile(new URL("../components/SessionSidebar-rows.tsx", import.meta.url), "utf8");
+  assert.match(rows, /api\/sessions\/\$\{encodeURIComponent\(session\.id\)\}\/archive/);
+  assert.match(rows, /sessionSidebar\.archiveLeafOnly/);
+  assert.match(rows, /sessionSidebar\.archiveConfirm/);
+  assert.match(source, /onOpenArchive/);
 });
 
 test("prompt controls preserve abort, steer, and follow-up RPC commands", async () => {
